requirements.txt
# REAEDP project dependencies (Python 3.8+)
# Install: pip install -r requirements.txt

numpy>=1.20
matplotlib>=3.3
Pillow>=9.0
pandas>=1.3
